predicate synthesis: simplify pub interface
We now synthesize one predicate per query. I've also added
a function for substituting a predicate into an ASTNode which
should be useful in future.

We should think about nice ways to reference-count the trees that
we're cloning: specifically, the fields of a Vec probably don't need
to be cloned, so we can use Rc<[Field]> instead?